requirements.txt
# Transcripción con Whisper
openai-whisper>=20231117

# PyTorch con soporte CUDA (instalar manualmente si es necesario)
# Para GPU NVIDIA: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchaudio>=2.0.0

# Diarización de hablantes
pyannote.audio>=3.1.0

# Generación de documentos Word
python-docx>=1.1.0

# Monitor de carpeta
watchdog>=4.0.0

# Procesamiento de video/audio
ffmpeg-python>=0.2.0

# Utilidades
tqdm>=4.65.0
